move Rect to rust
