+++ ja2lib/HelpScreen.c
@@ -5,7 +5,6 @@
 #include "HelpScreen.h"
 #include "HelpScreenText.h"
 #include "Laptop/Laptop.h"
-#include "Rect.h"
 #include "SGP/ButtonSystem.h"
 #include "SGP/CursorControl.h"
 #include "SGP/Debug.h"
@@ -32,6 +31,7 @@
 #include "Utils/Utilities.h"
 #include "Utils/WordWrap.h"
 #include "platform.h"
+#include "rust_geometry.h"
 
 extern INT16 gsVIEWPORT_END_Y;
 extern void PrintDate(void);
