--- ja2lib/Rect.h
@@ -1,13 +1,0 @@
-#ifndef __RECT_H
-#define __RECT_H
-
-#include <stdint.h>
-
-struct Rect {
-  int32_t left;
-  int32_t top;
-  int32_t right;
-  int32_t bottom;
-};
-
-#endif
