+++ platform-win32/Input_win.c
@@ -4,7 +4,6 @@
 
 #include "Local.h"
 #include "Point.h"
-#include "Rect.h"
 #include "SGP/Debug.h"
 #include "SGP/English.h"
 #include "SGP/Input.h"
