+++ rustlib/src/exp_geometry.rs
@@ -9,6 +9,16 @@ pub struct GRect {
     iBottom: i32,
 }
 
+#[repr(C)]
+// #[derive(Default, Copy, Clone)]
+/// Copy of Rect from the C codebase.
+pub struct Rect {
+    left: i32,
+    top: i32,
+    right: i32,
+    bottom: i32,
+}
+
 // struct Box {
 //     left: u16,
 //     top: u16,
@@ -18,11 +28,22 @@ pub struct GRect {
 
 #[no_mangle]
 /// Create new GRect structure
-pub extern "C" fn NewRect(left: i32, top: i32, width: i32, height: i32) -> GRect {
+pub extern "C" fn NewGRect(left: i32, top: i32, width: i32, height: i32) -> GRect {
     GRect {
         iLeft: left,
         iTop: top,
         iRight: left + width,
         iBottom: top + height,
+    }
+}
+
+#[no_mangle]
+/// Create new Rect structure
+pub extern "C" fn NewRect(left: i32, top: i32, width: i32, height: i32) -> Rect {
+    Rect {
+        left,
+        top,
+        right: left + width,
+        bottom: top + height,
     }
 }
